+++ .env.example
@@ -1,12 +1,17 @@
-# The port that probot will listen.
-PORT=4000
-
-# The ID of your GitHub App
-APP_ID=
-WEBHOOK_SECRET=development
-
 # Use `trace` to get verbose logging or `info` to show less
 LOG_LEVEL=debug
 
+# Server setup
+PORT=3000
+
+# The ID of your GitHub App
+APP_ID=0000
+
+# Webhook Secret of your Github App
+WEBHOOK_SECRET=secret
+
 # Go to https://smee.io/new set this to the URL that you are redirected to.
-WEBHOOK_PROXY_URL=
+WEBHOOK_PROXY_URL=https://smee.io/sKv2tq3YNtViwnN
+
+REDIS_URL='databasea_url' # Leave empty for local db
+PRIVATE_KEY=private_key
