+++ package.json
@@ -26,7 +26,7 @@
   },
   "dependencies": {
     "dotenv": "6.0.0",
-    "ioredis": "3.2.2",
+    "ioredis": "4.0.0",
     "lodash": "4.17.10",
     "probot": "7.1.0",
     "typedi": "0.8.0",
@@ -35,7 +35,7 @@
   "devDependencies": {
     "@types/node": "10.5.3",
     "@types/lodash": "4.14.113",
-    "@types/ioredis": "3.2.14",
+    "@types/ioredis": "4.0.1",
     "@types/jest": "23.3.1",
     "@octokit/webhooks": "5.0.0",
     "@taqtile/gulp-module-resolver": "0.2.2",
