+++ src/data/db/index.ts
@@ -1,1 +1,2 @@
+export * from './db.client';
 export * from './developer.datasource';
