--- src/data/db/redis-client.ts
@@ -1,3 +1,0 @@
-import * as Redis from 'ioredis';
-
-export const redisClient = new Redis(process.env.REDIS_URL);
