--- src/data/local/github/index.ts
@@ -1,1 +1,0 @@
-export * from './track-reader';
