--- src/data/mappers/index.ts
@@ -1,1 +1,0 @@
-export * from './github-events';
