--- src/domain/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './developer.model';
-export * from './track.model';
