--- src/domain/usecases/index.ts
@@ -1,1 +1,0 @@
-export * from './next-step.use-case';
