--- src/presentation/mappers/index.ts
@@ -1,1 +1,0 @@
-export * from './github-payload.mapper';
