+++ src/presentation/robot.ts
@@ -1,19 +1,20 @@
 import { Application } from 'probot';
 import { Container } from 'typedi';
-import { GithubEvents } from '@data/mappers/github-events';
-import { DeveloperInput, Track } from '@domain/entities';
-import { NextStepUseCase } from '@domain/usecases';
-import { GithubEventSender } from '@presentation/github-interaction';
-import { PayloadMapper } from '@presentation/mappers';
-
-const eventsSender: GithubEventSender = Container.get(GithubEventSender);
-const useCase: NextStepUseCase = Container.get(NextStepUseCase);
+import { DeveloperInput, NextStepUseCase, Track } from '@domain';
+import { GithubEventSender } from './events-sender.service';
+import { GithubEvents } from './github-events.constants';
+import { PayloadMapper } from './payload.mapper';
 
 export const robot = (app: Application) => {
+  const eventsSender: GithubEventSender = Container.get(GithubEventSender);
+  const useCase: NextStepUseCase = Container.get(NextStepUseCase);
 
   app.on(GithubEvents.Installation.Created, async context => {
+
     const dev: DeveloperInput = PayloadMapper.mapToDeveloper(context.payload);
+    console.log('Dev', dev);
     const track: Track = await useCase.execute(dev);
+    console.log('Track', track);
 
     eventsSender.openEvent(context, track);
   });
@@ -27,17 +28,6 @@ export const robot = (app: Application) => {
     const isFinishComment: boolean = comment && comment.body.match(/finish|next*/i);
 
     if (isFinishComment) {
-      const dev: DeveloperInput = PayloadMapper.mapToDeveloper(context.payload);
-      const track: Track = await useCase.execute(dev);
-
-      eventsSender.openEvent(context, track);
-    }
-  });
-
-  app.on(GithubEvents.Clone, async context => {
-    const isRightClone = true
-
-    if (isRightClone) {
       const dev: DeveloperInput = PayloadMapper.mapToDeveloper(context.payload);
       const track: Track = await useCase.execute(dev);
